+++ src/tools/Built-in/6_command/CSnippetCommand.ts
@@ -20,7 +20,8 @@ export class CSnippetCommand extends vscode.CompletionItem {
             ? body
                 .replaceAll(/\s*,\s*\[/gu, '[')
                 .replaceAll(/\[\s*,\s*/gu, '[')
-                .replaceAll(/\s*\[\s*/gu, ',[$0 ')
+                .replaceAll(/\s*\[\s*/gu, ',$0 ')
+                .replaceAll(/\s*\]\s*/gu, '')
             : body;
 
         this.insertText = new vscode.SnippetString(body2);