+++ src/tools/Built-in/7_sub_command/Menu/Menu.tools.ts
@@ -36,7 +36,8 @@ export const { snippetMenu, MenuMDMap } = (() => {
             ? body
                 .replaceAll(/\s*,\s*\[/gu, '[')
                 .replaceAll(/\[\s*,\s*/gu, '[')
-                .replaceAll(/\s*\[\s*/gu, ',[$0 ')
+                .replaceAll(/\s*\[\s*/gu, ',$0 ')
+                .replaceAll(/\s*\]\s*/gu, '')
             : body;
         item.insertText = new vscode.SnippetString(body2);
         item.detail = 'Menu';